# [MM-27293] Remove heap pre-allocation for file uploads