--- a/app/file.go
+++ b/app/file.go
@@ -545,8 +545,11 @@ type UploadFileTask struct {
 
 func (t *UploadFileTask) init(a *App) {
 	t.buf = &bytes.Buffer{}
-	t.maxFileSize = *a.Config().FileSettings.MaxFileSize
-	t.limit = *a.Config().FileSettings.MaxFileSize
+	if t.ContentLength > 0 {
+		t.limit = t.ContentLength
+	} else {
+		t.limit = t.maxFileSize
+	}
 
 	t.fileinfo = model.NewInfo(filepath.Base(t.Name))
 	t.fileinfo.Id = model.NewId()
@@ -554,19 +557,6 @@ func (t *UploadFileTask) init(a *App) {
 	t.fileinfo.CreateAt = t.Timestamp.UnixNano() / int64(time.Millisecond)
 	t.fileinfo.Path = t.pathPrefix() + t.Name
 
-	// Prepare to read ContentLength if it is known, otherwise limit
-	// ourselves to MaxFileSize. Add an extra byte to check and fail if the
-	// client sent too many bytes.
-	if t.ContentLength > 0 {
-		t.limit = t.ContentLength
-		// Over-Grow the buffer to prevent bytes.ReadFrom from doing it
-		// at the very end.
-		t.buf.Grow(int(t.limit + 1 + bytes.MinRead))
-	} else {
-		// If we don't know the upload size, grow the buffer somewhat
-		// anyway to avoid extra reslicing.
-		t.buf.Grow(UploadFileInitialBufferSize)
-	}
 	t.limitedInput = &io.LimitedReader{
 		R: t.Input,
 		N: t.limit + 1,
@@ -587,14 +577,14 @@ func (a *App) UploadFileX(channelId, name string, input io.Reader,
 	opts ...func(*UploadFileTask)) (*model.FileInfo, *model.AppError) {
 
 	t := &UploadFileTask{
-		ChannelId: filepath.Base(channelId),
-		Name:      filepath.Base(name),
-		Input:     input,
+		ChannelId:   filepath.Base(channelId),
+		Name:        filepath.Base(name),
+		Input:       input,
+		maxFileSize: *a.Config().FileSettings.MaxFileSize,
 	}
 	for _, o := range opts {
 		o(t)
 	}
-	t.init(a)
 
 	if len(*a.Config().FileSettings.DriverName) == 0 {
 		return nil, t.newAppError("api.file.upload_file.storage.app_error",
@@ -604,6 +594,8 @@ func (a *App) UploadFileX(channelId, name string, input io.Reader,
 		return nil, t.newAppError("api.file.upload_file.too_large_detailed.app_error",
 			"", http.StatusRequestEntityTooLarge, "Length", t.ContentLength, "Limit", t.maxFileSize)
 	}
+
+	t.init(a)
 
 	var aerr *model.AppError
 	if !t.Raw && t.fileinfo.IsImage() {